+++ src/pages/projContracts/parts/Preview/ContractDialog.tsx
@@ -6,6 +6,8 @@ import useDeepCompareEffect from 'use-deep-compare-effect';
 import CloseIcon from '@mui/icons-material/Close';
 import { SelectDocuments } from './SelectDocuments';
 import { Loading } from './Loading';
+import { StartContract } from './startContract';
+import { useBackdrop } from '../../../../hooks';
 
 
 export const ContractDialog = ({
@@ -14,7 +16,7 @@ export const ContractDialog = ({
   open: boolean,
   handleClose: () => void
 }) => {
-
+  const { backdropState: { open: backdropOpen } } = useBackdrop();
   const { values } = useFormikContext<TypeOfForm>();
 
   const {
@@ -30,6 +32,7 @@ export const ContractDialog = ({
 
   }, [values, open]);
 
+  const isBusy = backdropOpen || previewLoading;
 
   return (
     <Dialog
@@ -69,22 +72,22 @@ export const ContractDialog = ({
           p: 0,
         }}
       >
-        {!previewLoading &&
+        {!isBusy &&
         <embed
           src={previewUrl}
           width="100%"
           height='100%'
         />}
-        {previewLoading && <Loading />}
+        {isBusy && <Loading />}
       </DialogContent>
-      {!previewLoading &&
+      {!isBusy &&
         <DialogActions>
           <Button>
             ダウンロード
           </Button>
-          <Button>
-            送信
-          </Button>
+          {!previewLoading &&
+          <StartContract />}
+
         </DialogActions>}
 
     </Dialog>
