+++ src/pages/projContracts/parts/Preview/startContract/ChoiceButton.tsx
